Shifted common functions to service named commonService
